test: split null metadata test into serialization/business logic
